mas-agent/requirements.txt
paho-mqtt==2.1.0
requests==2.32.3
prometheus-client==0.20.0
pydantic==2.7.1
python-dotenv==1.0.1
structlog==24.2.0
 
